requirements.txt
whois
dnspython
requests
beautifulsoup4
snscrape